fix(content): Various gnome restaurant fixes (#1091)
* fix(content): Veg ball is made from potatoes, not tomatoes

* fix(content): Remove redundant if_close statements

* fix(content): Give first unfinished fruit batta when cooking, not second

* fix(content): Fix all gnome foods being called "gnome dish" on last step of cooking

* fix(content): Remove double dot in message

Tanner believes he did this by mistake.

* fix(content): Stop special casing vegetable batta

In osrs it acts just like all other gnome food.

* fix(content): Remove old comment

%eat_delay2 and %drink_delay don't seem to exist anymore.

* fix(content): Add missing opheldu for orange slices

Needed to make the pineapple punch cocktail.

* fix(content): Fix swapped premade/non-premade gnomebowl items

* fix(content): Fix cocktail job assignment logic

When $rand was 0, you would be told to make a Fruit Blast but the
$gnome_bar_progress variable remained at 7. So you actually didn't get
any job at all.

* fix(content): Remove comma in gnome waiter dialougue

Comma doesn't exist in osrs.

* fix(content): Add missing cocktail delete check

* fix(content): Remove some redundant if statements

* fix(content): Add pineapple to the Wizard Blizzard ingredient list

The code already checks for it, it was just missing from the guide book.

* fix(content): Improve missing utensils checks talking to Gianne

It now matches how it works in osrs during the tutorial.

* fix(content): Remove gnome cooking from level up messages

Currently we have no level requirement for gnome cooking.

* fix(content): Fix Pineapple Punch ingredient order

* fix(content): Change behaviour of cooking empty half baked gnome food

The previous behaviour was that the animation and sound played, but you
still got the "You can't cook that." message. Since we don't know how it
actually worked, fix it by simplifying the code.

* fix(content): Consistently use mes() for gnome cooking

Don't think we have any sources, so just be consistent.

* fix(content): Remove category_15 and rename gnome cooking categories

category_15 seems to be used for items are are supposed to be
unobtainable in osrs. This has no meaning in 2004scape so using it just
made the code wrong.

---------

Co-authored-by: henrik <>